fix: keep Boltz-2's two fp32 trunk islands under a bf16 trunk
Upstream's bf16-mixed is not blanket bf16; the sequence-stack attention takes
`z` in fp32 and the scan carry is fp32 (`pairformer.py:105-110`). Narrowing
both with the rest of the trunk cost 0.5776 A against upstream on a matched
tape; restoring them gives 0.2091 A for 0.8 s.

`s_init` is cast before `zeros_like` so the scan carry's dtype is decided
before the loop -- `lax.scan` requires carry type equality, so a cast inside
the body would not type-check.

+++ src/foldjax/models/boltz2/models/trunk_blocks/pairformer.py
@@ -182,7 +182,12 @@ def pairformer_layer_forward(
     s = s + attention_pair_bias_forward(
         params["attention"],
         s=s_normed,
-        z=z,
+        # `z.float()`, as Boltz spells it at pairformer.py:107. The pair track
+        # itself stays in the autocast dtype -- this widening is for the bias
+        # this attention builds from it, and nothing downstream sees it. `s`
+        # arrives float32 already, promoted where the stack's scan carry is
+        # created so the carry types agree.
+        z=z.astype(jnp.float32),
         mask=mask.astype(jnp.float32),
         k_in=s_normed,
         eps=eps,
